move actions and reducers
